+++ annbench/algo/proxy.py
@@ -1,10 +1,26 @@
-from .annoy import AnnoyANN
-from .ivfpq import IvfpqANN, Ivfpq4bitANN
-from .hnsw import HnswANN
-from .linear import LinearANN
-from .linear_gpu import LinearGpuANN
-from .ivfpq_gpu import IvfpqGpuANN
-from .scann import ScannANN
+import importlib.util
+
+def check_module(module_name):
+    # Return true only if the module has been installed
+    spec = importlib.util.find_spec(module_name)
+    if spec is None:  # The module hasn't been installed
+        return False
+    else:
+        return True
+
+if check_module("annoy"):
+    from .annoy import AnnoyANN
+
+if check_module("hnswlib"):
+    from .hnsw import HnswANN
+
+if check_module("faiss"):
+    from .faiss_cpu import LinearANN, IvfpqANN, Ivfpq4bitANN, PqANN, Pq4bitANN
+    from .faiss_gpu import LinearGpuANN, IvfpqGpuANN
+
+if check_module("scann"):
+    from .scann import ScannANN
+
 
 def instantiate_algorithm(name):
     """
@@ -30,6 +46,10 @@ def instantiate_algorithm(name):
         return IvfpqGpuANN()
     elif name == "ivfpq4bit":
         return Ivfpq4bitANN()
+    elif name =="pq":
+        return PqANN()
+    elif name =="pq4bit":
+        return Pq4bitANN()
     elif name == "scann":
         return ScannANN()
     else:
