+++ conf/algo/ivfpq4bit.yaml
@@ -5,8 +5,16 @@ name: ivfpq4bit
 param_index:
   - M: 16
     nlist: 100
+    quantizer: flat
   - M: 32
     nlist: 100
+    quantizer: flat
+  - M: 16
+    nlist: 100
+    quantizer: hnsw
+  - M: 32
+    nlist: 100
+    quantizer: hnsw
 
 param_query:
   nprobe:
