+++ conf/specialized_param/deep1m_ivfpq.yaml
@@ -3,5 +3,13 @@ algo:
   param_index:
     - M: 8
       nlist: 1000
+      quantizer: flat
     - M: 16
       nlist: 1000
+      quantizer: flat
+    - M: 8
+      nlist: 1000
+      quantizer: hnsw
+    - M: 16
+      nlist: 1000
+      quantizer: hnsw
